+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 * Use enum names in generated test functions when defined in `burstValues`.
 
+**Fixed**
+
+* Make the test class constructor with default arguments synthetic. Fixes JUnit5 compatibility.
+
 ## [2.9.0] *(2025-09-04)*
 [2.9.0]: https://github.com/cashapp/burst/releases/tag/2.9.0