+++ CHANGELOG.md
@@ -11,8 +11,8 @@
 
 * Use enum names in generated test functions when defined in `burstValues`.
 * Make the test class constructor with default arguments synthetic. Fixes JUnit5 compatibility.
+* Support AGP's built-in Kotlin.
 * Use the test instance’s class name and package for the test functions of non-final classes. Previously we were using the function’s enclosing class name.
-
 
 
 ## [2.9.0] *(2025-09-04)*